代码优化

+++ windy-feature/src/main/java/com/zj/feature/service/TestCaseService.java
@@ -3,11 +3,18 @@ package com.zj.feature.service;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.zj.common.uuid.UniqueIdService;
 import com.zj.common.model.PageSize;
+import com.zj.domain.entity.dto.feature.ExecutePointDto;
+import com.zj.domain.entity.dto.feature.FeatureInfoDto;
 import com.zj.domain.entity.dto.feature.TestCaseDto;
+import com.zj.domain.repository.feature.IExecutePointRepository;
+import com.zj.domain.repository.feature.IFeatureRepository;
 import com.zj.domain.repository.feature.ITestCaseRepository;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.stereotype.Service;
-import org.springframework.util.CollectionUtils;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * @author guyuelan
@@ -17,10 +24,14 @@ import org.springframework.util.CollectionUtils;
 public class TestCaseService {
   private final UniqueIdService uniqueIdService;
   private final ITestCaseRepository testCaseRepository;
+  private final IFeatureRepository featureRepository;
+  private final IExecutePointRepository executePointRepository;
 
-  public TestCaseService(UniqueIdService uniqueIdService, ITestCaseRepository testCaseRepository) {
+  public TestCaseService(UniqueIdService uniqueIdService, ITestCaseRepository testCaseRepository, IFeatureRepository featureRepository, IExecutePointRepository executePointRepository) {
     this.uniqueIdService = uniqueIdService;
     this.testCaseRepository = testCaseRepository;
+    this.featureRepository = featureRepository;
+    this.executePointRepository = executePointRepository;
   }
 
   public PageSize<TestCaseDto> getTestCaseList(String serviceId, Integer page, Integer pageSize) {
@@ -55,7 +66,22 @@ public class TestCaseService {
     return testCaseRepository.getTestCaseById(caseId);
   }
 
+  @Transactional
   public Boolean deleteTestCase(String caseId) {
-    return testCaseRepository.deleteTestCase(caseId);
+    List<FeatureInfoDto> featureList = featureRepository.queryFeatureList(caseId);
+    if (CollectionUtils.isEmpty(featureList)) {
+      return testCaseRepository.deleteTestCase(caseId);
+    }
+
+    List<String> featureIds = featureList.stream().map(FeatureInfoDto::getFeatureId).collect(Collectors.toList());
+    List<ExecutePointDto> executePoints = executePointRepository.getPointsByFeatureIds(featureIds);
+
+    boolean deleteFeatures = featureRepository.batchDeleteByFeatureId(featureIds);
+    if (CollectionUtils.isEmpty(executePoints)) {
+      return deleteFeatures && testCaseRepository.deleteTestCase(caseId);
+    }
+
+    boolean deletePoints = executePointRepository.deleteByFeatureIds(featureIds);
+    return deleteFeatures && deletePoints && testCaseRepository.deleteTestCase(caseId);
   }
 }
